Moved Tagspage to more appropriate space. Removed unused features. Added route
