+++ src/routes.tsx
@@ -12,6 +12,7 @@ export const privateRoutes = [
   "minndla/folders/preview/:folderId/:subFolderId",
   "minndla/folders/preview/:folderId/:subFolderId/:resourceId",
   "minndla/folders/:folderId",
+  "minndla/folders/tag/:tag",
   "minndla/arena",
   "minndla/admin",
   "minndla/admin/flags",
@@ -22,8 +23,6 @@ export const privateRoutes = [
   "minndla/arena/category/:categoryId/edit",
   "minndla/arena/category/:categoryId/topic/new",
   "minndla/arena/topic/:topicId",
-  "minndla/tags",
-  "minndla/tags/:tag",
   "minndla/subjects",
   "minndla/arena/notifications",
   "minndla/profile",
